requirements/base.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.7
asyncio==3.4.3
acrcloud==1.5
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.5
flower==0.9.3  # https://github.com/mher/flower
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
future==0.18.2
idna==2.9

importlib-metadata==1.6.1
kombu==4.6.10


pyee==7.0.2
pyppeteer==0.2.2
pyppeteer-stealth==1.0.0
pytz==2020.1
pyopenssl
redis==3.5.3
requests==2.23.0
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.46.1
urllib3==1.25.9
vine==1.3.0
websockets==8.1
whitenoise==5.1.0
zipp==3.1.0
